Smarter fill-in-the-blanks (#617)
This PR introduces display:flex to our repetoire for styling fill in the blanks. For now, its use is limited to Design Recipes only, as it addresses #228 . Immediately, this results in all blanks growing to fit the available horizontal space.

Over time, we can expand this use into other worksheets, replacing brittle @fitb{XYZ} with @fitb{flex} to show that a blank should extend as far as possible. This makes maintenance a lot easier in the long run, since blanks "know how to stretch themselves".