Looks like make does compare dir timestamps, but the OS (?check) updates a dir's timestamp only based on its immediate descendents. Have the initialization target call a small script dir-hoist-timestamp.sh to correct the timestamps of all dirs in {pathways, lessons}. This will make rules based on their timestamps work, and the massage-*.sh scripts don't get called often, and when they do, they don't need to worry about timestamps which is really make's job.
The initialization target shouldn't be an explicit prereq, since it
always needs to be done, and if it is a prereq, it forces a rule to
always fire, which we don't want especially for the dir updating rules
now that we have timestamp comparison for them. To this end, put the
initialization in its own make phase. There are now 3 Makefile.phaseN (N
= 1,2,3)