Fix screenshots (#2093)
This revises the screenshot-elts file for the deployment case, using puppeteer-cluster to bulk-generate the slide images for every lesson plan at once. On my machine it's more than a 6x improvement over the current behavior.

This also introduces the notion of a puppeteer cache, which gradually builds up local copies of resources needed by our web pages. Over time, this makes pages load a lot faster. Would be interesting to port this over to our PDF generation as well, and see if that can become faster.

In process of creating this, I found and resolved quite a few slowdowns in our page-loading speed! A big one was the lack of gzip compression on the server. A simple config change on hostinger led to pages loading almost 3x faster! However, we still get killed by loading lots of external resources. Some of them can be sped up thanks to <script defer>, which I've added for PDFlib and download.js. But the use of external fonts (Lato, MathJax and Hack) is really killing us. I've created a separate issue to track that, but that doesn't need to block this merge.