PDF stability improvements
We're still sensitive to race conditions when using >1 tab-per-core, so this commit drops us back down to 1. Of course, this kills our performance - nearly doubling the time it takes to generate PDFs! To make up some of that time, I went ahead and implemented a simple cache for all fonts, js, and css files. I also extended the blocked list to be more aggressive.

On balance, the cache gives us back most of the performance we lost by dropping down to 1 tab-per-core.