Don't turn on headers
@retabak You didn't make any mistakes here, but this is a great example of why "flexible" languages that try to be smart are just a source of endless pain.

AsciiDoctor tries to be cute, and one example of this is how you can declaring table formatting without DECLARING table formatting. Pop quiz!! For which of the following two tables have a header?

```
[cols="1,1", options="header"]
|===
| Heading A	| Heading B
| Row 1 cell1	| Row 2 cell2
|===
```

```
[cols="1,1"]
|===
| Heading A	| Heading B

| Row 1 cell1	| Row 2 cell2
|===
```

The answer is F#@KING BOTH OF THEM, because for some reason people think it's too hard to type `options="header"`. Of course this leads to all kinds of confusion, which is one of the reasons we're trying to stop putting blank lines in between rows!!

Anyway, what happened when you added the `@ifsoln{....}` wrappers is that the second line of the row becomes _empty_ when you're not in solutions mode. This, in turn, makes AsciiDoctor think you wanted a header row, which caused a whole cascade of other errors.

Of course, the use-case we have is totally natural! Sometimes the second line of the first row is a solution, and we shouldn't have to do acrobatic to get around AsciiDoctor's "cleverness". So rather than come up with something bizarre, I just relied on the CORRECT way of declaring table formatting, and used `options="noheader"`.