Integrate mdquiz (#2323) (#2494)
This PR adds a top-level command ./build-mdquiz, which scans all directories within distribution looking for "assessments" folders. For each one that it finds, it copies the necessary template files to that directory and builds a standalone quiz based on those files and the JSON configuration within.

This PR also converts all .toml files to .json, removing the older files along with some WSL .Zone-Identifier crap