Lesson units (closes #2478)
* Group related lessons into units in  narrative

lesson-order.txt optionally groups its lessons into units, with the unit
name given in [brackets]. These unit names are shown in the pathway
narrative page. #2497

CSS class .lesson-unit for styling.

* Skip unit name when collecting course's lessons

{collect-distribution-info,make-pathway-tocs,make-workbook-jsons}.lua,
when extracting unit lessons from the s-exp in .workbook-lessons.rkt.kp,
should lop off the unit name. #2497

* Preserve unit info in pathway-tocs.js

The lesson array associated with a course is not a flat list of lessons,
but a list of unitless lessons and objects mapping a unit name to a list
of its lesson. #2497

* slides: --skipnode and use new unitized lesson-list

- build-slide, build-all-slides have a --skipnode option that avoids
  calling md2googleslides (during debugging)
- build-all-slides updated to use the new unit-based
  .workbook-lessons.rkt.kp

* unitless lessons get a vacuous unit name NO_UNIT #2497

* update unit order for Algebra 2

* hide #_no_unit unit

* improve CSS formatting of unit titles, and adjust titles for ALg2

* more CSS improvements

---------

Co-authored-by: Dorai Sitaram <ds26gte@gmail.com>
Co-authored-by: Emmanuel Schanzer <schanzer@ccs.neu.edu>